Modified Recrystillization Hardness Model
Modified the recrystillization hardness  to make estimation more conservative (take the minimum reduction in hardness instead of the mean). This has the benefit of making the estimates in reduced hardness more precise. Also switched the sampling from numpyro to scipy.stats because this isn't really complicated enough to warrant using numpyro, and we're just sampling from independent t distributions. It's much faster now.