Latent Model Specification and Inference
Major progress on development of the latent model. Effectively allows fitting all recrystallization data sets using one model, which in turns allows "interpolation" between varying data sets using intuitive measures of recrystallization speed.

Major additions include the specification of the models required for this inference, as
well as some stochastic optimizers for maximum likelihood estimation vs. segregated optimization on KDE of marginal posteriors.

A significant amount of visualization is incorporated into a number of notebooks.

After some slight further analysis of the latent model and it's applications, this project will effectively be concluded. The major effort will then be in providing thorough documentation and succinct summary so that this work can be immediately useful to those that find it.